Split type analysis modules for lint
